requirements.txt
setuptools==75.1.0
wheel==0.44.0
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.6.1
joblib==1.3.1
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
zxing-cpp==3.0.0
pyzbar==0.1.9
gunicorn==23.0.0
